feat: provide `custom-inventory` input
Use this input to pass a list of particular VMs for the upgrade to run against. It can be used to
target VMs that were 'unreachable' in previous runs. It saves time by narrowing the scope of
machines to run against.

`testnet-deploy` will look up the IP addresses of the VMs named using the argument, then write
those IP addresses out to a custom inventory file.